requirements.txt
Flask==3.0.3
matplotlib==3.9.2
numpy==2.1.2
Pillow==10.4.0
scipy==1.14.1
sounddevice==0.5.0
torch==2.4.1
transformers==4.45.1
gradio==5.0.2
accelerate==1.0.1